+++ CHANGELOG.md
@@ -43,6 +43,8 @@ In addition to a few bug fixes and a bunch of new quality-of-life features (incl
 - **Armor Trims**
   - Hamster Armor can now be customized with vanilla armor trims in the Smithing Table.
   - The following trims are currently available: Coast, Border, Vex, Eye, Sentry, Wild
+  - They will naturally glow in the dark and automatically have emission with shader mods like Iris.
+  - Added an `Emissive Armor Trims` toggle in the client config so you can disable the glowing effect if you hate fun and prefer a non-luminescent reality.
 - **"Hamtaro" Easter Egg**
   - Renaming a hamster "Hamtaro" gives it a special texture. This functions identically to the "Sweet Potato" easter egg. Both base textures provided by [**@jimcerberus**](https://this_person_did_not_want_to_include_a_link_but_I_wanted_their_name_to_be_blue.com)!
 - **New "Panda" Overlay Pattern**